reorganize project to hide internals.
Public modules (those under Airship/ but outside Airship/Internal)
are now entirely exports modules. All actual code is stored in
Airship.Internal.*.
